+++ src/compas_libigl/parametrisation.py
@@ -5,7 +5,7 @@ from compas_libigl import _parametrisation
 
 
 @plugin(category="trimesh")
-def trimesh_harmonic(M):
+def trimesh_harmonic_mapping(M):
     """Compute the harmonic parametrisation of a triangle mesh within a fixed circular boundary.
 
     Parameters
@@ -19,9 +19,6 @@ def trimesh_harmonic(M):
     list[list[float]]
         The u, v parameters per vertex.
 
-    Notes
-    -----
-    ``camelhead.off`` can be downloaded from https://raw.githubusercontent.com/libigl/libigl-tutorial-data/master/camelhead.off
     """
     V, F = M
     V = np.asarray(V, dtype=np.float64)
@@ -30,7 +27,7 @@ def trimesh_harmonic(M):
 
 
 @plugin(category="trimesh")
-def trimesh_lscm(M):
+def trimesh_lsc_mapping(M):
     """Compute the least squares conformal map of a triangle mesh.
 
     Parameters
@@ -44,9 +41,6 @@ def trimesh_lscm(M):
     list[list[float]]
         The u, v parameters per vertex.
 
-    Notes
-    -----
-    ``camelhead.off`` can be downloaded from https://raw.githubusercontent.com/libigl/libigl-tutorial-data/master/camelhead.off
     """
     V, F = M
     V = np.asarray(V, dtype=np.float64)